Split eigensolvers into separate files, checkpointing
