+++ include/eigensolve_quda.h
@@ -7,7 +7,6 @@
 
 namespace quda
 {
-
   // Local enum for the LU axpy block type
   enum blockType { PENCIL, LOWER_TRI, UPPER_TRI };
 
@@ -148,11 +147,20 @@ public:
     void orthonormalizeGS(std::vector<ColorSpinorField *> &v, int j);
 
     /**
-       @brief Check orthonormality of input vector space v
+       @brief Orthonormalise input vector space v using Modified Gram-Schmidt
        @param[in] v Vector space
        @param[in] j Use vectors v[0:j-1] 
     */
-    void orthoCheck(std::vector<ColorSpinorField *> v, int j);
+    void orthonormalizeMGS(std::vector<ColorSpinorField *> &v, int j);
+
+    /**
+       @brief Check orthonormality of input vector space v
+       @param[out] bool If all vectors are orthonormal to 1e-16 returns true,
+       else false.
+       @param[in] v Vector space
+       @param[in] j Use vectors v[0:j-1] 
+    */
+    bool orthoCheck(std::vector<ColorSpinorField *> v, int j);
     
     /**
        @brief Permute the vector space using the permutation matrix.
@@ -337,9 +345,6 @@ public:
     double *alpha;
     double *beta;
 
-    // Used to clone vectors and resize arrays.
-    ColorSpinorParam csParam;
-
     /**
        @brief Compute eigenpairs
        @param[in] kSpace Krylov vector space
@@ -375,12 +380,13 @@ public:
 
   };
 
+
+  
   /**
      @brief Block Thick Restarted Lanczos Method.
   */
   class BLKTRLM : public TRLM
   {
-
   public:
     /**
        @brief Constructor for Thick Restarted Eigensolver class
@@ -389,56 +395,62 @@ public:
        @param profile Time Profile
     */
     BLKTRLM(const DiracMatrix &mat, QudaEigParam *eig_param, TimeProfile &profile);
-    
+      
     /**
        @brief Destructor for Thick Restarted Eigensolver class
     */
     virtual ~BLKTRLM();
-    
+      
     virtual bool hermitian() { return true; } /** (BLOCK)TRLM is only for Hermitian systems */
-    
+      
     // Variable size matrix
     std::vector<Complex> block_ritz_mat;
-    
+      
     /** Block Tridiagonal/Arrow matrix, fixed size. */
     Complex *block_alpha;
     Complex *block_beta;
-
+      
     /** Temp storage used in blockLanczosStep */
     Complex *jth_block;
     double *beta_diag_inv;
-    
+    double *alpha_old;
+      
     int n_blocks; //** Number of blocks in Krylov space */
     int block_data_length; //** Size of blocks of data in alpha/beta */
-    
+      
     /**
        @brief Compute eigenpairs
        @param[in] kSpace Krylov vector space
        @param[in] evals Computed eigenvalues
     */
     void operator()(std::vector<ColorSpinorField *> &kSpace, std::vector<Complex> &evals);
-
+      
     /**
        @brief block lanczos step: extends the Kylov space in block step
        @param[in] v Vector space
        @param[in] j Index of block of vectors being computed
     */
     void blockLanczosStep(std::vector<ColorSpinorField *> v, int j);    
-
+      
     /**
-       @brief Get the eigendecomposition from the block arrow matrix
-       @param[in] nLocked Number of locked eigenvectors
-       @param[in] arrow_pos position of arrowhead
+       @brief Get the eigendecomposition from the current block arrow matrix
+       @param[in] block_pos position of current block
     */
-    void eigensolveFromBlockArrowMat(int nLocked, int arror_pos);
-
+    void eigensolveFromBlockArrowMat(int block_pos);
+      
+    /**
+       @brief Accumulate the R products of QR into the block beta array
+       @param[in] k The QR iteration
+       @param[in] arrow_offset The current block position
+    */
+    void updateBlockBeta(int k, int arrow_offset);
+      
     /**
        @brief Rotate the Ritz vectors usinng the arrow matrix eigendecomposition
        Uses a complex ritz matrix
        @param[in] nKspace current Krylov space
     */
     void computeBlockKeptRitz(std::vector<ColorSpinorField *> &kSpace);
-    
   };
   
   /**
